+++ packages/react-preferences/src/components/ChannelPreferences.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { PreferenceItemComponentFn } from "./types";
+import { PreferenceItemComponentFn } from "../types";
 
 // TODO
 export const ChannelPreferences: PreferenceItemComponentFn = () => <></>;
