+++ packages/react-preferences/src/components/SnoozePreference.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { PreferenceItemComponentFn } from "./types";
+import { PreferenceItemComponentFn } from "../types";
 
 // TODO
 export const SnoozePreference: PreferenceItemComponentFn = () => <></>;
