--- packages/react-preferences/src/components/types.ts
@@ -1,15 +1,0 @@
-import {
-  PreferenceStatus,
-  ChannelClassification,
-  SnoozePreference,
-  IPreference,
-} from "~/hooks";
-
-export type Preference = "channel_preferences" | "status" | "snooze";
-
-export type PreferenceItemComponentFn = React.FunctionComponent<{
-  key: number;
-  label: string;
-  value: PreferenceStatus | ChannelClassification[] | SnoozePreference;
-  handleOnPreferenceChange: (changes: Partial<IPreference>) => void;
-}>;
